Now Cortex-A9 work with GCC 7.2
git-svn-id: http://dolphin/svn/svn2/trunk@39 953d4507-ba89-49e6-903b-66a71853bf49

+++ hardware_r7s721.h
@@ -14,6 +14,6 @@ void r7s721_intc_registintfunc(uint_fast16_t int_id, void (* func)(void));
 void r7s721_handler_interrupt(uint32_t icciar);
 
 void r7s721_ttb_initialize(void);
-void r7s721_ttb_map(unsigned long va,	/* virtual address */unsigned long la	/* linear (physical) address */);
+void r7s721_ttb_map(uintptr_t va,	/* virtual address */uintptr_t la	/* linear (physical) address */);
 
 #endif /* HARDWARE_R7S721_H_INCLUDED */
